Move old config loading logic to oldconfigs package.
This is a refactor step to preserve the legacy logic since we still need
to use this to check ConstraintTemplates until we migrate policy-library
to the new ConstraintTemplate format.
